requirements.txt
# Orcamentista IA — runtime Python (Vercel @vercel/python@5.x, Python 3.13)
# Bundle target: <250MB. Cold start: <2s.

# --- Web framework ---
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
pydantic>=2.10

# --- HTTP / utils ---
httpx>=0.27
python-dotenv>=1.0

# --- LLM vision (multi-provider) ---
# Sprint 1+ implementation. Stub-friendly: import on-demand inside routers.
gaik[vision-extract]>=0.5.11
markdown-it-py>=3.0

# --- DXF parser (Sprint 2) ---
ezdxf>=1.3

# --- Excel export (Sprint 4) ---
openpyxl>=3.1

# --- Supabase ---
supabase>=2.7

# --- Auth (JWT verification) ---
python-jose[cryptography]>=3.3

# --- SINAPI matcher (Sprint 3) ---
# 384-dim embeddings via sentence-transformers MiniLM-L6-v2.
# WARN: torch + sentence-transformers ~150MB. May force migration to OpenAI embeddings if Vercel bundle exceeds 250MB.
sentence-transformers>=2.6
numpy>=2.0
